Changed price to be decimal type instead of double

+++ server/FONdrum/FONdrum.Domain/Models/Wine.cs
@@ -6,7 +6,7 @@
 
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int StockQuantity { get; set; }
         public Guid StyleId { get; set; }
         public WineStyle Style { get; set; }
@@ -16,7 +16,7 @@
 
         public Wine() { }
 
-        public Wine(string name, double price, int stockQuantity, WineStyle style, GrapeVariety variety, string? imageUrl = null)
+        public Wine(string name, decimal price, int stockQuantity, WineStyle style, GrapeVariety variety, string? imageUrl = null)
         {
             Name = name;
             Price = price;
